+++ src/main/java/com/tuvarna/transportsystem/dao/UserTypeDAO.java
@@ -1,6 +1,7 @@
 package com.tuvarna.transportsystem.dao;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.function.Consumer;
 
 import javax.persistence.EntityManager;
@@ -39,16 +40,15 @@ public class UserTypeDAO implements GenericDAOInterface<UserType> {
 
 	@SuppressWarnings("unchecked")
 	@Override
-	public UserType getById(int id) {
-		return (UserType) entityManager.createQuery("FROM UserType WHERE usertype_id = :id").setParameter("id", id)
-				.getSingleResult(); // check if the return type has to be Optional<Class> or it
-									// is ok like this
+	public Optional<UserType> getById(int id) {
+		return Optional.ofNullable((UserType) entityManager.createQuery("FROM UserType WHERE usertype_id = :id").setParameter("id", id)
+				.getSingleResult()); 
 	}
 
 	@Override
-	public UserType getByName(String name) {
-		return (UserType) entityManager.createQuery("FROM UserType WHERE usertype_name = :name")
-				.setParameter("name", name).getSingleResult();
+	public Optional<UserType> getByName(String name) {
+		return Optional.ofNullable((UserType) entityManager.createQuery("FROM UserType WHERE usertype_name = :name")
+				.setParameter("name", name).getSingleResult());
 	}
 
 	@Override
@@ -70,13 +70,21 @@ public class UserTypeDAO implements GenericDAOInterface<UserType> {
 
 	@Override
 	public void deleteById(int id) {
-		UserType type = this.getById(id);
+		if (!this.getById(id).isPresent()) {
+			return;
+		}
+		
+		UserType type = this.getById(id).get();
 		executeInsideTransaction(entityManager -> entityManager.remove(type));
 	}
 
 	@Override
 	public void deleteByName(String name) {
-		UserType type = this.getByName(name);
+		if (!this.getByName(name).isPresent()) {
+			return;
+		}
+		
+		UserType type = this.getByName(name).get();
 		executeInsideTransaction(entityManager -> entityManager.remove(type));
 	}
